Add support for STI subclasses
STI subclasses were untaggable because the taggable type validation would see a mismatch between the tag.taggable_type and the value in the polymorphic taggable type column of the Tagging. This is because Rails stores the base class in polymorphic association "type" columns, not the STI subclass.

Instead, we strikes a balance between validity and performance by checking if the tag's allowed taggable type is a subclass of whatever taggable type is being stored in the tagging. We could have looked up the taggable to see if its class matched, but that would require a database query, which seems like overkill for what this is doing.